requirements.txt
fastapi
uvicorn[standard]
pydantic_settings
SQLAlchemy
pymysql
sqlalchemy[asyncio] 
aiomysql
python-jose
cryptography
langchain_google_genai